# docs/openapi/masterdata.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
paths:
  /api/v1/masterdata/wards/uid/{uid}:
    get:
      tags:
      - ward-controller
      operationId: getByUid
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WardDto"
    put:
      tags:
      - ward-controller
      operationId: update
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WardRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WardDto"
  /api/v1/masterdata/ward-types/uid/{uid}:
    get:
      tags:
      - ward-type-controller
      operationId: getByUid_1
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WardTypeDto"
    put:
      tags:
      - ward-type-controller
      operationId: update_1
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WardTypeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WardTypeDto"
  /api/v1/masterdata/ward-categories/uid/{uid}:
    get:
      tags:
      - ward-category-controller
      operationId: getByUid_2
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WardCategoryDto"
    put:
      tags:
      - ward-category-controller
      operationId: update_2
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WardCategoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WardCategoryDto"
  /api/v1/masterdata/theatres/uid/{uid}:
    get:
      tags:
      - theatre-controller
      operationId: getByUid_3
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TheatreDto"
    put:
      tags:
      - theatre-controller
      operationId: update_3
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TheatreRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TheatreDto"
  /api/v1/masterdata/suppliers/uid/{uid}:
    get:
      tags:
      - supplier-controller
      operationId: getByUid_4
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SupplierDto"
    put:
      tags:
      - supplier-controller
      operationId: update_4
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SupplierDto"
  /api/v1/masterdata/supplier-item-prices/uid/{uid}:
    get:
      tags:
      - supplier-item-price-controller
      operationId: getByUid_5
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SupplierItemPriceDto"
    put:
      tags:
      - supplier-item-price-controller
      operationId: update_5
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierItemPriceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SupplierItemPriceDto"
    delete:
      tags:
      - supplier-item-price-controller
      operationId: delete
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/v1/masterdata/stores/uid/{uid}:
    get:
      tags:
      - store-controller
      operationId: getByUid_6
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StoreDto"
    put:
      tags:
      - store-controller
      operationId: update_6
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StoreDto"
  /api/v1/masterdata/radiology-types/uid/{uid}:
    get:
      tags:
      - radiology-type-controller
      operationId: getByUid_7
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RadiologyTypeDto"
    put:
      tags:
      - radiology-type-controller
      operationId: update_7
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RadiologyTypeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RadiologyTypeDto"
  /api/v1/masterdata/procedure-types/uid/{uid}:
    get:
      tags:
      - procedure-type-controller
      operationId: getByUid_8
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProcedureTypeDto"
    put:
      tags:
      - procedure-type-controller
      operationId: update_8
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcedureTypeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProcedureTypeDto"
  /api/v1/masterdata/pharmacies/uid/{uid}:
    get:
      tags:
      - pharmacy-controller
      operationId: getByUid_9
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PharmacyDto"
    put:
      tags:
      - pharmacy-controller
      operationId: update_9
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PharmacyRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PharmacyDto"
  /api/v1/masterdata/medicines/uid/{uid}:
    get:
      tags:
      - medicine-controller
      operationId: getByUid_10
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MedicineDto"
    put:
      tags:
      - medicine-controller
      operationId: update_10
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MedicineDto"
  /api/v1/masterdata/lab-test-types/uid/{uid}:
    get:
      tags:
      - lab-test-type-controller
      operationId: getByUid_11
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LabTestTypeDto"
    put:
      tags:
      - lab-test-type-controller
      operationId: update_11
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabTestTypeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LabTestTypeDto"
...
  /api/v1/masterdata/items/uid/{uid}:
    get:
      tags:
      - item-controller
      operationId: getByUid_12
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ItemDto"
    put:
      tags:
      - item-controller
      operationId: update_12
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ItemDto"
  /api/v1/masterdata/item-suppliers/uid/{uid}:
    get:
      tags:
      - item-supplier-controller
      operationId: getByUid_13
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ItemSupplierDto"
    put:
      tags:
      - item-supplier-controller
      operationId: update_13
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemSupplierRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ItemSupplierDto"
  /api/v1/masterdata/item-medicine-coefficients/uid/{uid}:
    get:
      tags:
      - item-medicine-coefficient-controller
      operationId: getByUid_14
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ItemMedicineCoefficientDto"
    put:
      tags:
      - item-medicine-coefficient-controller
      operationId: update_14
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemMedicineCoefficientRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ItemMedicineCoefficientDto"
  /api/v1/masterdata/insurance-providers/uid/{uid}:
    get:
      tags:
      - insurance-provider-controller
      operationId: getByUid_15
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InsuranceProviderDto"
    put:
      tags:
      - insurance-provider-controller
      operationId: update_15
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsuranceProviderRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InsuranceProviderDto"
  /api/v1/masterdata/insurance-plans/uid/{uid}:
    get:
      tags:
      - insurance-plan-controller
      operationId: getByUid_16
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InsurancePlanDto"
    put:
      tags:
      - insurance-plan-controller
      operationId: update_16
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsurancePlanRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InsurancePlanDto"
  /api/v1/masterdata/diagnosis-types/uid/{uid}:
    get:
      tags:
      - diagnosis-type-controller
      operationId: getByUid_17
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DiagnosisTypeDto"
    put:
      tags:
      - diagnosis-type-controller
      operationId: update_17
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosisTypeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DiagnosisTypeDto"
...
  /api/v1/masterdata/clinics/uid/{uid}:
    get:
      tags:
      - clinic-controller
      operationId: getByUid_18
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClinicDto"
    put:
      tags:
      - clinic-controller
      operationId: update_19
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClinicDto"
  /api/v1/masterdata/beds/uid/{uid}:
    get:
      tags:
      - ward-bed-controller
      operationId: getByUid_19
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WardBedDto"
    put:
      tags:
      - ward-bed-controller
      operationId: update_20
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WardBedRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WardBedDto"
...
  /api/v1/masterdata/radiology-types:
    get:
      tags:
      - radiology-type-controller
      operationId: list_8
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RadiologyTypeDto"
    post:
      tags:
      - radiology-type-controller
      operationId: create_7
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RadiologyTypeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RadiologyTypeDto"
...
components:
  schemas:
    WardRequest:
      required:
      - code
      - name
      - wardCategoryUid
      - wardTypeUid
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        noOfBeds:
          type: integer
          format: int32
        active:
          type: boolean
        wardCategoryUid:
          type: string
        wardTypeUid:
          type: string
    WardDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        name:
          type: string
        noOfBeds:
          type: integer
          format: int32
        active:
          type: boolean
        wardCategoryUid:
          type: string
        wardTypeUid:
          type: string
    WardTypeRequest:
      required:
      - code
      - name
      - price
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        active:
          type: boolean
    WardTypeDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        active:
          type: boolean
    WardCategoryRequest:
      required:
      - code
      - name
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
    WardCategoryDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
    TheatreRequest:
      required:
      - code
      - name
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        location:
          type: string
        active:
          type: boolean
    TheatreDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        location:
          type: string
        active:
          type: boolean
    SupplierRequest:
      required:
      - code
      - contactName
      - name
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        contactName:
          type: string
        active:
          type: boolean
        tin:
          type: string
        vrn:
          type: string
        termsOfContract:
          type: string
        physicalAddress:
          type: string
        postCode:
          type: string
        postAddress:
          type: string
        telephone:
          type: string
        mobile:
          type: string
        email:
          type: string
        fax:
          type: string
        bankAccountName:
          type: string
        bankPhysicalAddress:
          type: string
        bankPostCode:
          type: string
        bankPostAddress:
          type: string
        bankName:
          type: string
        bankAccountNo:
          type: string
    SupplierDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        name:
          type: string
        contactName:
          type: string
        active:
          type: boolean
        tin:
          type: string
        vrn:
          type: string
        termsOfContract:
          type: string
        physicalAddress:
          type: string
        postCode:
          type: string
        postAddress:
          type: string
        telephone:
          type: string
        mobile:
          type: string
        email:
          type: string
        fax:
          type: string
        bankAccountName:
          type: string
        bankPhysicalAddress:
          type: string
        bankPostCode:
          type: string
        bankPostAddress:
          type: string
        bankName:
          type: string
        bankAccountNo:
          type: string
    SupplierItemPriceRequest:
      required:
      - itemUid
      - price
      - supplierUid
      type: object
      properties:
        price:
          type: number
        terms:
          type: string
        active:
          type: boolean
        supplierUid:
          type: string
        itemUid:
          type: string
    SupplierItemPriceDto:
      type: object
      properties:
        uid:
          type: string
        price:
          type: number
        terms:
          type: string
        active:
          type: boolean
        supplierUid:
          type: string
        itemUid:
          type: string
    StoreRequest:
      required:
      - code
      - name
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        location:
          type: string
        category:
          type: string
        active:
          type: boolean
    StoreDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        location:
          type: string
        category:
          type: string
        active:
          type: boolean
    RadiologyTypeRequest:
      required:
      - code
      - name
      - price
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        uom:
          type: string
        active:
          type: boolean
    RadiologyTypeDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        uom:
          type: string
        active:
          type: boolean
    ProcedureTypeRequest:
      required:
      - code
      - name
      - price
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        uom:
          type: string
        active:
          type: boolean
    ProcedureTypeDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        uom:
          type: string
        active:
          type: boolean
    PharmacyRequest:
      required:
      - code
      - name
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        location:
          type: string
        category:
          type: string
        active:
          type: boolean
    PharmacyDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        location:
          type: string
        category:
          type: string
        active:
          type: boolean
    MedicineRequest:
      required:
      - category
      - code
      - name
      - price
      - type
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
        price:
          type: number
        uom:
          type: string
        category:
          type: string
        active:
          type: boolean
    MedicineDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
        price:
          type: number
        uom:
          type: string
        category:
          type: string
        active:
          type: boolean
    LabTestTypeRequest:
      required:
      - code
      - name
      - price
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        uom:
          type: string
        active:
          type: boolean
    LabTestTypeDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        uom:
          type: string
        active:
          type: boolean
...
    ItemRequest:
      required:
      - code
      - costPriceVatIncl
      - name
      - packSize
      - sellingPriceVatIncl
      - vat
      type: object
      properties:
        code:
          type: string
        barcode:
          type: string
        name:
          type: string
        shortName:
          type: string
        commonName:
          type: string
        vat:
          type: number
        uom:
          type: string
        packSize:
          type: number
        category:
          type: string
        costPriceVatIncl:
          type: number
        sellingPriceVatIncl:
          type: number
        active:
          type: boolean
        ingredients:
          type: string
    ItemDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        barcode:
          type: string
        name:
          type: string
        shortName:
          type: string
        commonName:
          type: string
        vat:
          type: number
        uom:
          type: string
        packSize:
          type: number
        category:
          type: string
        costPriceVatIncl:
          type: number
        sellingPriceVatIncl:
          type: number
        active:
          type: boolean
        ingredients:
          type: string
    ItemSupplierRequest:
      required:
      - costPriceVatExcl
      - costPriceVatIncl
      - itemUid
      - supplierUid
      type: object
      properties:
        itemUid:
          type: string
        supplierUid:
          type: string
        costPriceVatIncl:
          type: number
        costPriceVatExcl:
          type: number
        active:
          type: boolean
    ItemSupplierDto:
      type: object
      properties:
        uid:
          type: string
        itemUid:
          type: string
        supplierUid:
          type: string
        costPriceVatIncl:
          type: number
        costPriceVatExcl:
          type: number
        active:
          type: boolean
    ItemMedicineCoefficientRequest:
      required:
      - itemQty
      - itemUid
      - medicineQty
      - medicineUid
      type: object
      properties:
        itemUid:
          type: string
        medicineUid:
          type: string
        itemQty:
          type: number
        medicineQty:
          type: number
    ItemMedicineCoefficientDto:
      type: object
      properties:
        uid:
          type: string
        itemUid:
          type: string
        medicineUid:
          type: string
        coefficient:
          type: number
        itemQty:
          type: number
        medicineQty:
          type: number
    InsuranceProviderRequest:
      required:
      - code
      - name
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        address:
          type: string
        telephone:
          type: string
        email:
          type: string
        fax:
          type: string
        website:
          type: string
        active:
          type: boolean
    InsuranceProviderDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        name:
          type: string
        address:
          type: string
        telephone:
          type: string
        email:
          type: string
        fax:
          type: string
        website:
          type: string
        active:
          type: boolean
...
    InsurancePlanDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
        insuranceProviderUid:
          type: string
...
    DiagnosisTypeDto:
      type: object
      properties:
        uid:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
...